Added Code to Randomly Set Wild Card as a Dummy Color Card
In the multiplayer game, if a Wild or Wild Plus Four card is the first card placed onto the discard pile (before the player or any of the CPUs places a card), the game automatically changes it into a dummy color card. This avoids the issue where the players must play another Wild or Wild Plus Four card onto the starting Wild card.